# swagger.json
{
	"openapi": "3.1.0",
	"info": {
	  "title": "GiftReactApp",
	  "contact": {},
	  "version": "1.0"
	},
...
	"servers": [
	  {
		"url": "http://example.com",
		"variables": {}
	  }
	],
	"paths": {
...
	  "/upload": {
		"post": {
		  "tags": [
			"Misc"
		  ],
		  "summary": "FileUpload",
		  "operationId": "FileUpload",
		  "parameters": [
			{
			  "name": "Content-Type",
			  "in": "header",
			  "description": "",
			  "required": true,
			  "style": "simple",
			  "schema": {
				"const": "application/json",
				"type": "string",
				"examples": [
				  "application/json"
				]
			  }
			}
		  ],
		  "requestBody": {
			"content": {
			  "application/x-www-form-urlencoded": {
				"encoding": {},
				"schema": {
				  "required": [
					"file"
				  ],
				  "type": "object",
				  "properties": {
					"file": {
					  "type": "string"
					}
				  },
				  "contentMediaType": "application/x-www-form-urlencoded"
				}
			  }
			},
			"required": false
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {}
			}
		  },
		  "deprecated": false,
		  "security": [
			{
			  "bearer": []
			}
		  ]
		},
		"parameters": []
	  },
...
	},
	"components": {
...
	  "securitySchemes": {
		"bearer": {
		  "type": "http",
		  "scheme": "bearer"
		}
	  }
	},
	"security": [],
...
  }